requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
python-dotenv==1.0.0
numpy==1.24.3
PyPDF2==3.0.1
Pillow==10.0.0
pdf2image==1.16.3
python-multipart==0.0.6 